Safari-611.1.10/LayoutTests/inspector/unit-tests/collection-expected.txt

Testing all methods of Collection.


== Running test suite: Collection
-- Running test case: WI.Collection.prototype.add
["one","two"]
["one","two",3]

-- Running test case: WI.Collection.prototype.remove
["one","two"]
["two"]

-- Running test case: WI.Collection.prototype.clear
["one","two",3]
[]

-- Running test case: WI.Collection.SpecifiedType
["one","two"]
["one","two"]
